+++ .gitignore
@@ -1,5 +1,4 @@
 # Created by .ignore support plugin (hsz.mobi)
-sonosco/pycandle/
 experiments/
 tests/test_wavs/
