+++ requirements.txt
@@ -23,3 +23,4 @@ wget==3.2
 pytest==4.6.3
 click==7.0
 deprecation==2.0.6
+dataclasses==0.6
