+++ sonosco/datasets/__init__.py
@@ -1,3 +1,3 @@
-from .audio_dataset import AudioDataProcessor, AudioDataset
-from .data_sampler import BucketingSampler
-from .data_loader import AudioDataLoader, create_data_loaders
+from .dataset import AudioDataProcessor, AudioDataset
+from .samplers import BucketingSampler
+from .loader import AudioDataLoader, create_data_loaders
